Tally face centered

+++ movement/watch_faces/complication/tally_face.h
@@ -38,7 +38,7 @@
 #include "movement.h"
 
 typedef struct {
-    uint32_t tally_idx;
+    int16_t tally_idx;
     uint8_t tally_default_idx : 7;
     bool soundOff;
 } tally_state_t;
